services/model-executor/requirements.txt
# gRPC and protobuf
grpcio==1.64.1
grpcio-tools==1.64.1
protobuf==5.27.1

# Real ML inference
sentence-transformers==3.0.1
torch==2.3.1                   # CPU build; swap for cu121 variant with GPU nodes
transformers==4.42.4
huggingface-hub==0.24.0

# Observability
prometheus-client==0.20.0

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
